fix: Add top-level `text` argument in the slack message : (#128)
```
The top-level `text` argument is missing in the request payload for a chat.postMessage call - It's a best practice to always provide a `text` argument when posting a message. The `text` is used in places where the content cannot be rendered such as: system push notifications, assistive technology such as screen readers, etc.
```

Co-authored-by: Thibaut FARCIN <thibaut.farcin@manomano.com>